requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.6
django-environ==0.9.0
django-extensions==3.2.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
mssql-django==1.2
mysqlclient==2.1.1
pycparser==2.21
PyMySQL==1.0.2
pyodbc==4.0.35
pytz==2022.7.1
sqlparse==0.4.3
